@@ -1,7 +1,9 @@
+from pathlib import Path
+
+from sphinx_js import renderers
+
 from .jsdoc import (
     patch_sphinx_js,
-    ts_post_convert,
-    ts_should_destructure_arg,
     ts_xref_formatter,
 )
 from .lexers import HtmlPyodideLexer, PyodideLexer
@@ -55,7 +57,13 @@ def setup(app):
     app.setup_extension("sphinx_js")
     app.add_directive("pyodide-package-list", get_packages_summary_directive(app))
     app.connect("builder-inited", add_mdn_xrefs)
-    app.config.ts_post_convert = ts_post_convert
-    app.config.ts_should_destructure_arg = ts_should_destructure_arg
     app.config.ts_type_xref_formatter = ts_xref_formatter
     app.config.ts_type_bold = True
+    app.config.ts_sphinx_js_config = Path(__file__).parent / "sphinxJsConfig.ts"
+    renderers._SECTION_ORDER = [
+        "type_aliases",
+        "interfaces",
+        "attributes",
+        "functions",
+        "classes",
+    ]
