@@ -1052,7 +1052,7 @@ done:
       dt->shared = H5FL_FREE(H5T_shared_t, dt->shared);
       dt = H5FL_FREE(H5T_t, dt);
     } /* end if */
-  }   /* end if */
+  } /* end if */
 
   FUNC_LEAVE_NOAPI(ret_value);
 } /* end H5T__init_native() */
