@@ -2,14 +2,16 @@ import pytest
 from pytest_pyodide import run_in_pyodide
 
 
-@pytest.mark.driver_timeout(40)
+@pytest.mark.xfail_browsers(firefox="Too slow")
 @run_in_pyodide(packages=["sisl-tests", "pytest"])
-def test_version(selenium):
-    import sisl
+def test_load_sisl(selenium):
+    """Loading sisl takes a really long time so this separates it out to reduce
+    the chance that test_nodes times out.
+    """
+    pass
 
-    assert sisl.__version__ == "0.14.2"
 
-
+@pytest.mark.xfail_browsers(firefox="Too slow")
 @run_in_pyodide(packages=["sisl-tests", "pytest"])
 def test_nodes(selenium):
     import pytest
@@ -17,6 +19,7 @@ def test_nodes(selenium):
     pytest.main(["--pyargs", "sisl.nodes"])
 
 
+@pytest.mark.xfail_browsers(firefox="Too slow")
 @run_in_pyodide(packages=["sisl-tests", "pytest"])
 def test_geom(selenium):
     import pytest
@@ -24,6 +27,7 @@ def test_geom(selenium):
     pytest.main(["--pyargs", "sisl.geom"])
 
 
+@pytest.mark.xfail_browsers(firefox="Too slow")
 @run_in_pyodide(packages=["sisl-tests", "pytest"])
 def test_linalg(selenium):
     import pytest
@@ -31,6 +35,7 @@ def test_linalg(selenium):
     pytest.main(["--pyargs", "sisl.linalg"])
 
 
+@pytest.mark.xfail_browsers(firefox="Too slow")
 @run_in_pyodide(packages=["sisl-tests", "pytest"])
 def test_sparse(selenium):
     import pytest
@@ -38,6 +43,7 @@ def test_sparse(selenium):
     pytest.main(["--pyargs", "sisl.tests.test_sparse"])
 
 
+@pytest.mark.xfail_browsers(firefox="Too slow")
 @run_in_pyodide(packages=["sisl-tests", "pytest"])
 def test_physics_sparse(selenium):
     import pytest
