@@ -13,7 +13,6 @@ static PyObject* tbmod = NULL;
 static PyObject* _pyodide_importhook = NULL;
 
 _Py_IDENTIFIER(__qualname__);
-_Py_IDENTIFIER(format_exception);
 _Py_IDENTIFIER(add_note_to_module_not_found_error);
 
 void
@@ -62,54 +61,15 @@ set_error(PyObject* err)
 EM_JS(
 JsVal,
 new_error,
-(const char* type, const char* msg, PyObject* err),
+(const char* type, JsVal msg, PyObject* err),
 {
-  return new API.PythonError(UTF8ToString(type), UTF8ToString(msg), err);
+  return new API.PythonError(UTF8ToString(type), msg, err);
 });
 // clang-format on
 
 /**
- * Fetch the exception, normalize it, and ensure that traceback is not NULL.
- *
- * Always succeeds, always results in type, value, traceback not NULL.
- */
-static void
-fetch_and_normalize_exception(PyObject** type,
-                              PyObject** value,
-                              PyObject** traceback)
-{
-  PyErr_Fetch(type, value, traceback);
-  PyErr_NormalizeException(type, value, traceback);
-  if (*type == NULL || Py_IsNone(*type) || *value == NULL ||
-      Py_IsNone(*value)) {
-    Py_CLEAR(*type);
-    Py_CLEAR(*value);
-    Py_CLEAR(*traceback);
-    fail_test();
-    PyErr_SetString(PyExc_TypeError,
-                    "Pyodide internal error: no exception type or value");
-    PyErr_Fetch(type, value, traceback);
-    PyErr_NormalizeException(type, value, traceback);
-  }
-
-  if (*traceback == NULL) {
-    *traceback = Py_None;
-    Py_INCREF(*traceback);
-  }
-  PyException_SetTraceback(*value, *traceback);
-}
-
-static void
-store_sys_last_exception(PyObject* type, PyObject* value, PyObject* traceback)
-{
-  PySys_SetObject("last_type", type);
-  PySys_SetObject("last_value", value);
-  PySys_SetObject("last_traceback", traceback);
-}
-
-/**
- * Restore sys.last_exception as the current exception if sys.last_value matches
- * the argument value. Used for reentrant errors.
+ * Restore sys.last_exception as the current exception if sys.last_exc matches
+ * the argument `exc`. Used for reentrant errors.
  * Returns true if it restored the error indicator, false otherwise.
  *
  * If we throw a JavaScript PythonError and it bubbles out to the enclosing
@@ -124,67 +84,48 @@ store_sys_last_exception(PyObject* type, PyObject* value, PyObject* traceback)
  * support for catching errors by type.
  */
 EMSCRIPTEN_KEEPALIVE bool
-restore_sys_last_exception(void* value)
+restore_sys_last_exception(void* exc)
 {
-  bool success = false;
-  FAIL_IF_NULL(value);
-  PyObject* last_type = PySys_GetObject("last_type");
-  FAIL_IF_NULL(last_type);
-  PyObject* last_value = PySys_GetObject("last_value");
-  FAIL_IF_NULL(last_value);
-  PyObject* last_traceback = PySys_GetObject("last_traceback");
-  FAIL_IF_NULL(last_traceback);
-  if (value != last_value) {
-    return 0;
+  if (exc == NULL) {
+    return false;
   }
-  // PyErr_Restore steals a reference to each of its arguments so need to incref
-  // them first.
-  Py_INCREF(last_type);
-  Py_INCREF(last_value);
-  Py_INCREF(last_traceback);
-  PyErr_Restore(last_type, last_value, last_traceback);
-  success = true;
-finally:
-  return success;
+  // PySys_GetObject returns a borrowed reference and will return NULL without
+  // setting an exception if it fails.
+  PyObject* last_exc = PySys_GetObject("last_exc");
+  if (last_exc != exc) {
+    return false;
+  }
+  // PyErr_SetRaisedException steals a reference to its argument and
+  // PySys_GetObject returns a borrow so need to incref last_xxc first.
+  Py_INCREF(last_exc);
+  PyErr_SetRaisedException(last_exc);
+  return true;
 }
 
-EM_JS(void, fail_test, (), { API.fail_test = true; })
+// clang-format off
+EM_JS(void, fail_test, (), {
+  API.fail_test = true;
+})
 
-/**
- * Calls traceback.format_exception(type, value, traceback) and joins the
- * resulting list of strings together.
- */
-static PyObject*
-format_exception_traceback(PyObject* type, PyObject* value, PyObject* traceback)
-{
-  PyObject* pylines = NULL;
-  PyObject* empty = NULL;
-  PyObject* result = NULL;
+EM_JS(void, capture_stderr, (void), {
+  API.capture_stderr();
+});
 
-  pylines = _PyObject_CallMethodIdObjArgs(
-    tbmod, &PyId_format_exception, type, value, traceback, NULL);
-  FAIL_IF_NULL(pylines);
-  empty = PyUnicode_New(0, 0);
-  FAIL_IF_NULL(empty);
-  result = PyUnicode_Join(empty, pylines);
-  FAIL_IF_NULL(result);
-
-finally:
-  Py_CLEAR(pylines);
-  Py_CLEAR(empty);
-  return result;
-}
+EM_JS(JsVal, restore_stderr, (void), {
+  return API.restore_stderr();
+});
+// clang-format on
 
 /**
  * Wrap the exception in a JavaScript PythonError object.
  *
- * The return value of this function is always a valid hiwire ID to an error
- * object. It never returns NULL.
+ * The return value of this function is always a JavaScript error object. It
+ * never returns null.
  *
  * We are cautious about leaking the Python stack frame, so we don't increment
  * the reference count on the exception object, we just store a pointer to it.
- * Later we can check if this pointer is equal to sys.last_value and if so
- * restore the exception (see restore_sys_last_exception).
+ * Later we can check if this pointer is equal to sys.last_exc and if so restore
+ * the exception (see restore_sys_last_exception).
  *
  * WARNING: dereferencing the error pointer stored on the PythonError is a
  * use-after-free bug.
@@ -193,29 +134,44 @@ EMSCRIPTEN_KEEPALIVE JsVal
 wrap_exception()
 {
   bool success = false;
-  PyObject* type = NULL;
-  PyObject* value = NULL;
-  PyObject* traceback = NULL;
+  PyObject* exc = NULL;
   PyObject* typestr = NULL;
-  PyObject* pystr = NULL;
-  fetch_and_normalize_exception(&type, &value, &traceback);
-  store_sys_last_exception(type, value, traceback);
-  if (type == PyExc_ModuleNotFoundError) {
+
+  exc = PyErr_GetRaisedException();
+
+  if (PyErr_GivenExceptionMatches(exc, PyExc_ModuleNotFoundError)) {
     PyObject* res = _PyObject_CallMethodIdOneArg(
-      _pyodide_importhook, &PyId_add_note_to_module_not_found_error, value);
+      _pyodide_importhook, &PyId_add_note_to_module_not_found_error, exc);
     FAIL_IF_NULL(res);
     Py_CLEAR(res);
   }
 
-  typestr = _PyObject_GetAttrId(type, &PyId___qualname__);
+  capture_stderr();
+  PyErr_SetRaisedException(Py_NewRef(exc));
+  // print standard traceback to standard error, clear the error flag, and set
+  // sys.last_exc, sys.last_type, etc
+  //
+  // Calls sys.excepthook. We set the excepthook to call
+  // traceback.print_exception, see `set_excepthook()` in
+  // `_pyodide/__init__.py`.
+  //
+  // If the error is a SystemExit and the PyConfig.inspect flag is not set,
+  // PyErr_Print() will call exit(). We don't want this generally, so we will
+  // generally set the `inspect` flag. The exception is in the CLI runner.
+  //
+  // In the CLI runner, if we call back into JS then back into Python and the
+  // inner Python raises SystemExit, we won't actually unwind the Python frames
+  // in the outer Python. Hypothetically this could cause trouble and we should
+  // fix it, but it's probably not worth the effort.
+  PyErr_Print();
+  JsVal formatted_exception = restore_stderr();
+
+  typestr = _PyObject_GetAttrId((PyObject*)Py_TYPE(exc), &PyId___qualname__);
   FAIL_IF_NULL(typestr);
   const char* typestr_utf8 = PyUnicode_AsUTF8(typestr);
   FAIL_IF_NULL(typestr_utf8);
-  pystr = format_exception_traceback(type, value, traceback);
-  FAIL_IF_NULL(pystr);
-  const char* pystr_utf8 = PyUnicode_AsUTF8(pystr);
-  FAIL_IF_NULL(pystr_utf8);
-  JsVal jserror = new_error(typestr_utf8, pystr_utf8, value);
+
+  JsVal jserror = new_error(typestr_utf8, formatted_exception, exc);
   FAIL_IF_JS_NULL(jserror);
 
   success = true;
@@ -225,17 +181,15 @@ finally:
     PySys_WriteStderr(
       "Pyodide: Internal error occurred while formatting traceback:\n");
     PyErr_Print();
-    if (type != NULL) {
+    if (exc != NULL) {
       PySys_WriteStderr("\nOriginal exception was:\n");
-      PyErr_Display(type, value, traceback);
+      PyErr_DisplayException(exc);
     }
-    jserror = new_error(
-      "PyodideInternalError", "Error occurred while formatting traceback", 0);
+    Js_static_string(msg, "Error occurred while formatting traceback");
+    jserror = new_error("PyodideInternalError", JsvString_FromId(&msg), 0);
   }
-  Py_CLEAR(type);
-  Py_CLEAR(value);
-  Py_CLEAR(traceback);
-  Py_CLEAR(pystr);
+  Py_CLEAR(exc);
+  Py_CLEAR(typestr);
   return jserror;
 }
 
