@@ -279,7 +279,10 @@ EM_JS_VAL(JsVal, JsvObject_Values, (JsVal obj), {
 
 EM_JS_VAL(JsVal,
 JsvObject_toString, (JsVal obj), {
-  return obj.toString();
+  if (hasMethod(obj, "toString")) {
+    return obj.toString();
+  }
+  return Object.prototype.toString.call(obj);
 });
 
 
@@ -366,37 +369,11 @@ EM_JS_VAL(JsVal, JsvPromise_Resolve, (JsVal obj), {
   // clang-format on
 });
 
-// Either syncifyHandler will get filled in by stack_switching/suspenders.mjs or
-// stack switching is not available so syncify will always return an error in
-// JsProxy.c and syncifyHandler will never be called.
-EMSCRIPTEN_KEEPALIVE JsVal (*syncifyHandler)(JsVal promise) = NULL;
-
-EM_JS(void, JsvPromise_Syncify_handleError, (void), {
-  if (!Module.syncify_error) {
-    // In this case we tried to syncify in a context where there is no
-    // suspender. JsProxy.c checks for this case and sets the error flag
-    // appropriately.
-    return;
-  }
-  Module.handle_js_error(Module.syncify_error);
-  delete Module.syncify_error;
-})
-
-JsVal
-JsvPromise_Syncify(JsVal promise)
-{
-  JsVal result = syncifyHandler(promise);
-  if (JsvNull_Check(result)) {
-    JsvPromise_Syncify_handleError();
-  }
-  return result;
-}
-
 // ==================== Buffers ====================
 
 // clang-format off
 EM_JS_NUM(errcode, jslib_init_buffers_js, (), {
-  const dtypes_str = ["b", "B", "h", "H", "i", "I", "f", "d"].join(
+  const dtypes_str = Array.from("bBhHiIqQfd").join(
     String.fromCharCode(0)
   );
   const dtypes_ptr = stringToNewUTF8(dtypes_str);
@@ -414,6 +391,8 @@ EM_JS_NUM(errcode, jslib_init_buffers_js, (), {
     ["Uint32Array", [dtypes_map["I"], 4, true]],
     ["Float32Array", [dtypes_map["f"], 4, true]],
     ["Float64Array", [dtypes_map["d"], 8, true]],
+    ["BigInt64Array", [dtypes_map["q"], 8, true]],
+    ["BigUint64Array", [dtypes_map["Q"], 8, true]],
     // These last two default to Uint8. They have checked : false to allow use
     // with other types.
     ["DataView", [dtypes_map["B"], 1, false]],
@@ -505,6 +484,10 @@ MAKE_OPERATOR(greater_than_equal, >=);
 // clang-format off
 EM_JS_VAL(JsVal, JsvMap_New, (), {
   return new Map();
+})
+
+EM_JS_VAL(JsVal, JsvLiteralMap_New, (), {
+  return new API.LiteralMap();
 })
 // clang-format on
 
