@@ -9,21 +9,32 @@ import {
   loadLockFile,
 } from "./compat";
 
-import { createModule, initializeFileSystem, preloadWasm } from "./module";
+import { createSettings } from "./emscripten-settings";
 import { version } from "./version";
 
 import type { PyodideInterface } from "./api.js";
-import type { TypedArray, API, Module } from "./types";
-import type { PackageData } from "./load-package";
+import type { TypedArray, Module, PackageData } from "./types";
+import type { EmscriptenSettings } from "./emscripten-settings";
+import type { SnapshotConfig } from "./snapshot";
 export type { PyodideInterface, TypedArray };
 
 export { version, type PackageData };
 
-declare function _createPyodideModule(Module: any): Promise<void>;
+declare function _createPyodideModule(
+  settings: EmscriptenSettings,
+): Promise<Module>;
+
+// BUILD_ID is generated from hashing together pyodide.asm.js and
+// pyodide.asm.wasm in esbuild.config.outer.mjs
+//
+// It is used to check that memory snapshots were generated by the same build of
+// the runtime that is trying to use them. Attempting to use a snapshot from a
+// different build will fail badly. See logic in snapshot.ts.
+declare const BUILD_ID: string;
 
 /**
  * See documentation for loadPyodide.
- * @private
+ * @hidden
  */
 export type ConfigType = {
   indexURL: string;
@@ -39,6 +50,10 @@ export type ConfigType = {
   _node_mounts: string[];
   env: { [key: string]: string };
   packages: string[];
+  _makeSnapshot: boolean;
+  enableRunUntilComplete: boolean;
+  checkAPIVersion: boolean;
+  BUILD_ID: string;
 };
 
 /**
@@ -153,15 +168,38 @@ export async function loadPyodide(
      */
     packages?: string[];
     /**
-     * Opt into the old behavior where PyProxy.toString calls `repr` and not
-     * `str`.
+     * Opt into the old behavior where :js:func:`PyProxy.toString() <pyodide.ffi.PyProxy.toString>`
+     * calls :py:func:`repr` and not :py:class:`str() <str>`.
      * @deprecated
      */
     pyproxyToStringRepr?: boolean;
     /**
+     * Make loop.run_until_complete() function correctly using stack switching
+     */
+    enableRunUntilComplete?: boolean;
+    /**
+     * If true (default), throw an error if the version of Pyodide core does not
+     * match the version of the Pyodide js package.
+     */
+    checkAPIVersion?: boolean;
+    /**
+     * Used by the cli runner. If we want to detect a virtual environment from
+     * the host file system, it needs to be visible from when `main()` is
+     * called. The directories in this list will be mounted at the same address
+     * into the Emscripten file system so that virtual environments work in the
+     * cli runner.
      * @ignore
      */
     _node_mounts?: string[];
+    /** @ignore */
+    _makeSnapshot?: boolean;
+    /** @ignore */
+    _loadSnapshot?:
+      | Uint8Array
+      | ArrayBuffer
+      | PromiseLike<Uint8Array | ArrayBuffer>;
+    /** @ignore */
+    _snapshotDeserializer?: (obj: any) => any;
   } = {},
 ): Promise<PyodideInterface> {
   await initNodeModules();
@@ -182,29 +220,24 @@ export async function loadPyodide(
     env: {},
     packageCacheDir: indexURL,
     packages: [],
+    enableRunUntilComplete: false,
+    checkAPIVersion: true,
+    BUILD_ID,
   };
   const config = Object.assign(default_config, options) as ConfigType;
-  if (!config.env.HOME) {
-    config.env.HOME = "/home/pyodide";
-  }
-
-  const Module = createModule();
-  Module.print = config.stdout;
-  Module.printErr = config.stderr;
-  Module.arguments = config.args;
-
-  const API = { config } as API;
-  Module.API = API;
+  config.env.HOME ??= "/home/pyodide";
+  /**
+   * `PyErr_Print()` will call `exit()` if the exception is a `SystemError`.
+   * This shuts down the Python interpreter, which is a change in behavior from
+   * what happened before. In order to avoid this, we set the `inspect` config
+   * parameter which prevents `PyErr_Print()` from calling `exit()`. Except in
+   * the cli runner, we actually do want to exit. So set default to true and in
+   * cli runner we explicitly set it to false.
+   */
+  config.env.PYTHONINSPECT ??= "1";
+  const emscriptenSettings = createSettings(config);
+  const API = emscriptenSettings.API;
   API.lockFilePromise = loadLockFile(config.lockFileURL);
-
-  preloadWasm(Module, indexURL);
-  initializeFileSystem(Module, config);
-
-  const moduleLoaded = new Promise((r) => (Module.postRun = r));
-
-  // locateFile tells Emscripten where to find the data files that initialize
-  // the file system.
-  Module.locateFile = (path: string) => config.indexURL + path;
 
   // If the pyodide.asm.js script has been imported, we can skip the dynamic import
   // Users can then do a static import of the script in environments where
@@ -214,61 +247,61 @@ export async function loadPyodide(
     await loadScript(scriptSrc);
   }
 
+  let snapshot: Uint8Array | undefined = undefined;
+  if (options._loadSnapshot) {
+    const snp = await options._loadSnapshot;
+    if (ArrayBuffer.isView(snp)) {
+      snapshot = snp;
+    } else {
+      snapshot = new Uint8Array(snp);
+    }
+    emscriptenSettings.noInitialRun = true;
+    // @ts-ignore
+    emscriptenSettings.INITIAL_MEMORY = snapshot.length;
+  }
+
   // _createPyodideModule is specified in the Makefile by the linker flag:
   // `-s EXPORT_NAME="'_createPyodideModule'"`
-  await _createPyodideModule(Module);
-
-  // There is some work to be done between the module being "ready" and postRun
-  // being called.
-  await moduleLoaded;
+  const Module = await _createPyodideModule(emscriptenSettings);
   // Handle early exit
-  if (Module.exited) {
-    throw Module.exited.toThrow;
+  if (emscriptenSettings.exited) {
+    throw emscriptenSettings.exited.toThrow;
   }
   if (options.pyproxyToStringRepr) {
     API.setPyProxyToStringMethod(true);
   }
 
-  if (API.version !== version) {
-    throw new Error(
-      `\
+  if (API.version !== version && config.checkAPIVersion) {
+    throw new Error(`\
 Pyodide version does not match: '${version}' <==> '${API.version}'. \
 If you updated the Pyodide version, make sure you also updated the 'indexURL' parameter passed to loadPyodide.\
-`,
-    );
+`);
   }
   // Disable further loading of Emscripten file_packager stuff.
   Module.locateFile = (path: string) => {
     throw new Error("Didn't expect to load any more file_packager files!");
   };
 
-  const pyodide = API.finalizeBootstrap();
+  let snapshotConfig: SnapshotConfig | undefined = undefined;
+  if (snapshot) {
+    snapshotConfig = API.restoreSnapshot(snapshot);
+  }
+  // runPython works starting after the call to finalizeBootstrap.
+  const pyodide = API.finalizeBootstrap(
+    snapshotConfig,
+    options._snapshotDeserializer,
+  );
+  API.sys.path.insert(0, API.config.env.HOME);
 
-  // runPython works starting here.
   if (!pyodide.version.includes("dev")) {
     // Currently only used in Node to download packages the first time they are
     // loaded. But in other cases it's harmless.
     API.setCdnUrl(`https://cdn.jsdelivr.net/pyodide/v${pyodide.version}/full/`);
   }
+  API._pyodide.set_excepthook();
   await API.packageIndexReady;
-
-  let importhook = API._pyodide._importhook;
-  importhook.register_module_not_found_hook(
-    API._import_name_to_package_name,
-    API.lockfile_unvendored_stdlibs_and_test,
-  );
-
-  if (API.lockfile_info.version !== version) {
-    throw new Error(
-      "Lock file version doesn't match Pyodide version.\n" +
-        `   lockfile version: ${API.lockfile_info.version}\n` +
-        `   pyodide  version: ${version}`,
-    );
-  }
-  API.package_loader.init_loaded_packages();
-  if (config.fullStdLib) {
-    await pyodide.loadPackage(API.lockfile_unvendored_stdlibs);
-  }
+  // I think we want this initializeStreams call to happen after
+  // packageIndexReady? I don't remember why.
   API.initializeStreams(config.stdin, config.stdout, config.stderr);
   return pyodide;
 }
