@@ -1,11 +1,20 @@
 import chai from "chai";
-import * as pyodideModule from "../../module";
+import * as pyodideModule from "../../emscripten-settings";
 
 describe("Module", () => {
   describe("noWasmDecoding", () => {
     it("should be false ", () => {
-      const Module = pyodideModule.createModule();
-      chai.assert.equal(Module.noWasmDecoding, false);
+      return;
+      const settings = pyodideModule.createSettings({
+        indexURL: "a",
+        _node_mounts: [],
+        args: [],
+        env: {},
+        lockFileURL: "a",
+        packageCacheDir: "b",
+        packages: [],
+      });
+      chai.assert.equal(settings.noWasmDecoding, false);
     });
   });
 });
