@@ -3,7 +3,8 @@ import type { PyProxy, PyAwaitable } from "generated/pyproxy";
 import { type PyodideInterface } from "./api";
 import { type ConfigType } from "./pyodide";
 import { type InFuncType } from "./streams";
-import { type PackageData, type InternalPackageData } from "./load-package";
+import { SnapshotConfig } from "./snapshot";
+import { ResolvablePromise } from "./common/resolveable";
 
 export type TypedArray =
   | Int8Array
@@ -62,6 +63,7 @@ declare global {
   export const _PyMem_Free: (ptr: number) => number;
   export const _PyGILState_Check: () => number;
   export const __PyErr_CheckSignals: () => number;
+  export const _PyErr_SetRaisedException: (ptr: number) => void;
 }
 
 // Our own functions we use from JavaScript. These all need to be labeled with
@@ -95,17 +97,29 @@ declare global {
         ) => any),
   ) => any;
 
-  export const _pyproxy_getflags: (ptr: number) => number;
+  export const _pyproxy_getflags: (
+    ptr: number,
+    is_json_adaptor: boolean,
+  ) => number;
   export const __pyproxy_type: (ptr: number) => string;
   export const __pyproxy_repr: (ptr: number) => string;
-  export const __pyproxy_getitem: (obj: number, key: any) => any;
+  export const __pyproxy_getitem: (
+    obj: number,
+    key: any,
+    cache: Map<string, any>,
+    is_json_adaptor: boolean,
+  ) => any;
   export const __pyproxy_setitem: (ptr: number, key: any, value: any) => number;
   export const __pyproxy_delitem: (ptr: number, key: any) => number;
   export const __pyproxy_contains: (ptr: number, key: any) => number;
   export const __pyproxy_GetIter: (ptr: number) => number;
   export const __pyproxy_GetAIter: (ptr: number) => number;
   export const __pyproxy_aiter_next: (ptr: number) => any;
-  export const __pyproxy_iter_next: (ptr: number) => any;
+  export const __pyproxy_iter_next: (
+    ptr: number,
+    cache: Map<string, any>,
+    is_json_adaptor: boolean,
+  ) => any;
   export const __pyproxyGen_Send: (
     ptr: number,
     arg: any,
@@ -161,6 +175,7 @@ export type FSNode = {
   timestamp: number;
   rdev: number;
   contents: Uint8Array;
+  mode: number;
 };
 
 export type FSStream = {
@@ -211,9 +226,15 @@ export interface FS {
   mkdev: (path: string, dev: number) => FSNode;
   filesystems: any;
   stat: (path: string, dontFollow?: boolean) => any;
-  readdir: (node: FSNode) => string[];
+  readdir: (path: string) => string[];
   isDir: (mode: number) => boolean;
-  lookupPath: (path: string) => { node: FSNode };
+  isMountpoint: (mode: FSNode) => boolean;
+  lookupPath: (
+    path: string,
+    options?: {
+      follow_mount?: boolean;
+    },
+  ) => { node: FSNode };
   isFile: (mode: number) => boolean;
   writeFile: (path: string, contents: any, o?: { canOwn?: boolean }) => void;
   chmod: (path: string, mode: number) => void;
@@ -230,37 +251,86 @@ export interface FS {
   close: (stream: FSStream) => void;
   ErrnoError: { new (errno: number): Error };
   registerDevice<T>(dev: number, ops: FSStreamOpsGen<T>): void;
+  syncfs(dir: boolean, oncomplete: (val: void) => void): void;
+  findObject(a: string, dontResolveLastLink?: boolean): any;
+  readFile(a: string): Uint8Array;
+}
+
+/** @private */
+export type PreRunFunc = (Module: Module) => void;
+
+export type ReadFileType = (path: string) => Uint8Array;
+
+// File System-like type which can be passed to
+// Module.loadDynamicLibrary or Module.loadWebAssemblyModule
+export type LoadDynlibFS = {
+  readFile: ReadFileType;
+  findObject: (path: string, dontResolveLastLink: boolean) => any;
+};
+
+type DSO = any;
+
+export interface LDSO {
+  loadedLibsByName: {
+    [key: string]: DSO;
+  };
 }
 
 export interface Module {
-  noImageDecoding: boolean;
-  noAudioDecoding: boolean;
-  noWasmDecoding: boolean;
-  quit: (status: number, toThrow: Error) => void;
-  preRun: { (): void }[];
-  print?: (a: string) => void;
-  printErr?: (a: string) => void;
-  arguments: string[];
   API: API;
-  postRun: ((a: Module) => void) | ((a: Module) => void)[];
   locateFile: (file: string) => string;
   exited?: { toThrow: any };
   ENV: { [key: string]: string };
   PATH: any;
   TTY: any;
   FS: FS;
+  LDSO: LDSO;
   canvas?: HTMLCanvasElement;
-  addRunDependency: (id: string) => void;
-  removeRunDependency: (id: string) => void;
-  reportUndefinedSymbols: () => void;
+  addRunDependency(id: string): void;
+  removeRunDependency(id: string): void;
+  reportUndefinedSymbols(): void;
+  loadDynamicLibrary(
+    lib: string,
+    options?: {
+      loadAsync?: boolean;
+      nodelete?: boolean;
+      allowUndefined?: boolean;
+      global?: boolean;
+      fs: LoadDynlibFS;
+    },
+    localScope?: object | null,
+    handle?: number,
+  ): void;
+  getDylinkMetadata(binary: Uint8Array | WebAssembly.Module): {
+    neededDynlibs: string[];
+  };
+
   ERRNO_CODES: { [k: string]: number };
-  instantiateWasm?: (
-    imports: { [key: string]: any },
-    successCallback: (
-      instance: WebAssembly.Instance,
-      module: WebAssembly.Module,
-    ) => void,
-  ) => void;
+  stringToNewUTF8(x: string): number;
+  _compat_to_string_repr: number;
+  js2python_convert: (
+    obj: any,
+    options: {
+      depth?: number;
+      defaultConverter?: (
+        value: any,
+        converter: (value: any) => any,
+        cacheConversion: (input: any, output: any) => void,
+      ) => any;
+    },
+  ) => any;
+  _PropagatePythonError: typeof Error;
+  _Py_EMSCRIPTEN_SIGNAL_HANDLING: number;
+  Py_EmscriptenSignalBuffer: TypedArray;
+  HEAP8: Uint8Array;
+  __hiwire_get(a: number): any;
+  __hiwire_set(a: number, b: any): void;
+  __hiwire_immortal_add(a: any): void;
+  _jslib_init(): number;
+  _init_pyodide_proxy(): number;
+  jsWrapperTag: any; // Should be WebAssembly.Tag
+  getExceptionMessage(e: number): [string, string];
+  handle_js_error(e: any): void;
 }
 
 type LockfileInfo = {
@@ -270,9 +340,52 @@ type LockfileInfo = {
   python: string;
 };
 
-type Lockfile = {
+export type Lockfile = {
   info: LockfileInfo;
-  packages: Record<string, PackageData>;
+  packages: Record<string, InternalPackageData>;
+};
+
+export type PackageType =
+  | "package"
+  | "cpython_module"
+  | "shared_library"
+  | "static_library";
+
+// Package data inside pyodide-lock.json
+
+export interface PackageData {
+  name: string;
+  version: string;
+  fileName: string;
+  /** @experimental */
+  packageType: PackageType;
+}
+
+/**
+ * @hidden
+ */
+export type InternalPackageData = {
+  name: string;
+  version: string;
+  file_name: string;
+  package_type: PackageType;
+  install_dir: string;
+  sha256: string;
+  imports: string[];
+  depends: string[];
+};
+
+/**
+ * @hidden
+ */
+export type PackageLoadMetadata = {
+  name: string;
+  normalizedName: string;
+  channel: string;
+  depends: string[];
+  done: ResolvablePromise;
+  installPromise?: Promise<void>;
+  packageData: InternalPackageData;
 };
 
 export interface API {
@@ -316,6 +429,8 @@ export interface API {
   runPythonInternal_dict: any;
   saveState: () => any;
   restoreState: (state: any) => void;
+  scheduleCallback: (callback: () => void, timeout: number) => void;
+  detectEnvironment: () => Record<string, boolean>;
 
   package_loader: any;
   importlib: any;
@@ -343,13 +458,51 @@ export interface API {
     pkg: InternalPackageData,
     dynlibPaths: string[],
   ) => Promise<void>;
-
+  recursiveDependencies: (
+    names: string[],
+    errorCallback: (err: string) => void,
+  ) => Map<string, PackageLoadMetadata>;
   _Comlink: any;
 
   dsodir: string;
   sys: PyProxy;
   os: PyProxy;
 
-  finalizeBootstrap: () => PyodideInterface;
+  restoreSnapshot(snapshot: Uint8Array): SnapshotConfig;
+  makeSnapshot(serializer?: (obj: any) => any): Uint8Array;
+  saveSnapshot(): Uint8Array;
+  finalizeBootstrap: (
+    fromSnapshot?: SnapshotConfig,
+    snapshotDeserializer?: (obj: any) => any,
+  ) => PyodideInterface;
+  syncUpSnapshotLoad3(conf: SnapshotConfig): void;
+  abortSignalAny: (signals: AbortSignal[]) => AbortSignal;
   version: string;
+
+  LiteralMap: any;
 }
+
+// Subset of the API and Module that the package manager needs
+/**
+ * @hidden
+ */
+export type PackageManagerAPI = Pick<
+  API,
+  | "importlib"
+  | "package_loader"
+  | "lockfile_packages"
+  | "bootstrapFinalizedPromise"
+  | "sitepackages"
+  | "defaultLdLibraryPath"
+> & {
+  config: Pick<ConfigType, "indexURL" | "packageCacheDir">;
+};
+/**
+ * @hidden
+ */
+export type PackageManagerModule = Pick<
+  Module,
+  "reportUndefinedSymbols" | "PATH" | "loadDynamicLibrary" | "LDSO"
+> & {
+  FS: Pick<FS, "readdir" | "lookupPath" | "isDir" | "findObject" | "readFile">;
+};
