@@ -11,6 +11,7 @@ import linecache
 import tokenize
 from collections.abc import Generator
 from copy import deepcopy
+from importlib import import_module
 from io import StringIO
 from textwrap import dedent
 from types import CodeType
@@ -132,6 +133,8 @@ def _parse_and_compile_gen(
     mode: str = "exec",
     filename: str = "<exec>",
     flags: int = 0x0,
+    dont_inherit: bool = False,
+    optimize: int = -1,
 ) -> Generator[ast.Module, ast.Module, CodeType]:
     """Parse ``source``, then yield the AST, then compile the AST and return the
     code object.
@@ -160,7 +163,7 @@ def _parse_and_compile_gen(
         _last_expr_to_raise(mod)
 
     ast.fix_missing_locations(mod)
-    return compile(mod, filename, mode, flags=flags)
+    return compile(mod, filename, mode, flags, dont_inherit, optimize)
 
 
 ReturnMode = Literal["last_expr", "last_expr_or_assign", "none"]
@@ -214,6 +217,16 @@ class CodeRunner:
         The flags to compile with. See the documentation for the built-in
         :external:py:func:`compile` function.
 
+    dont_inherit :
+
+        Whether to inherit ``__future__`` imports from the outer code.
+        See the documentation for the built-in :external:py:func:`compile` function.
+
+    optimize :
+
+        Specifies the optimization level of the compiler. See the documentation
+        for the built-in :external:py:func:`compile` function.
+
     Examples
     --------
     >>> source = "1 + 1"
@@ -254,6 +267,8 @@ class CodeRunner:
         quiet_trailing_semicolon: bool = True,
         filename: str = "<exec>",
         flags: int = 0x0,
+        dont_inherit: bool = False,
+        optimize: int = -1,
     ):
         self._compiled = False
         self._source = source
@@ -264,6 +279,8 @@ class CodeRunner:
             quiet_trailing_semicolon=quiet_trailing_semicolon,
             filename=filename,
             flags=flags,
+            dont_inherit=dont_inherit,
+            optimize=optimize,
         )
         self.ast = next(self._gen)
 
@@ -407,6 +424,8 @@ def eval_code(
     quiet_trailing_semicolon: bool = True,
     filename: str = "<exec>",
     flags: int = 0x0,
+    dont_inherit: bool = False,
+    optimize: int = -1,
 ) -> Any:
     """Runs a string as Python source code.
 
@@ -497,6 +516,8 @@ def eval_code(
             quiet_trailing_semicolon=quiet_trailing_semicolon,
             filename=filename,
             flags=flags,
+            dont_inherit=dont_inherit,
+            optimize=optimize,
         )
         .compile()
         .run(globals, locals)
@@ -512,6 +533,8 @@ async def eval_code_async(
     quiet_trailing_semicolon: bool = True,
     filename: str = "<exec>",
     flags: int = 0x0,
+    dont_inherit: bool = False,
+    optimize: int = -1,
 ) -> Any:
     """Runs a code string asynchronously.
 
@@ -577,10 +600,20 @@ async def eval_code_async(
             quiet_trailing_semicolon=quiet_trailing_semicolon,
             filename=filename,
             flags=flags,
+            dont_inherit=dont_inherit,
+            optimize=optimize,
         )
         .compile()
         .run_async(globals, locals)
     )
+
+
+def _add_prefixes(s: set[str], mod: str) -> None:
+    [current, *rest] = mod.split(".")
+    s.add(current)
+    for part in rest:
+        current += f".{part}"
+        s.add(current)
 
 
 def find_imports(source: str) -> list[str]:
@@ -594,14 +627,18 @@ def find_imports(source: str) -> list[str]:
 
     Returns
     -------
-        A list of module names that are imported in ``source``. If ``source`` is not
-        syntactically correct Python code (after dedenting), returns an empty list.
+        A list of module names that are imported in ``source``. If ``source`` is
+        not syntactically correct Python code (after dedenting), returns an
+        empty list.
+
+        Given `import package.module`, `find_imports` will include both
+        `"package"` and `"package.module"` in the result.
 
     Examples
     --------
     >>> source = "import numpy as np; import scipy.stats"
     >>> find_imports(source)
-    ['numpy', 'scipy']
+    ['numpy', 'scipy', 'scipy.stats']
     """
     # handle mis-indented input from multi-line strings
     source = dedent(source)
@@ -610,17 +647,17 @@ def find_imports(source: str) -> list[str]:
         mod = ast.parse(source)
     except SyntaxError:
         return []
-    imports = set()
+    imports: set[str] = set()
     for node in ast.walk(mod):
         if isinstance(node, ast.Import):
             for name in node.names:
                 node_name = name.name
-                imports.add(node_name.split(".")[0])
+                _add_prefixes(imports, node_name)
         elif isinstance(node, ast.ImportFrom):
             module_name = node.module
             if module_name is None:
                 continue
-            imports.add(module_name.split(".")[0])
+            _add_prefixes(imports, module_name)
     return list(sorted(imports))
 
 
@@ -628,5 +665,8 @@ def pyimport_impl(path: str) -> Any:
     [stem, *fromlist] = path.rsplit(".", 1)
     res = __import__(stem, fromlist=fromlist)
     if fromlist:
-        res = getattr(res, fromlist[0])
+        try:
+            res = getattr(res, fromlist[0])
+        except AttributeError:
+            res = import_module(path)
     return res
