@@ -286,6 +286,13 @@ def test_pyproxy_iter_error2(selenium):
     )
 
 
+@run_in_pyodide
+def test_pyproxy_iter_leak(selenium):
+    from js import Map
+
+    Map.new([(1, 2), (3, 4)])
+
+
 def test_pyproxy_get_buffer(selenium):
     selenium.run_js(
         """
@@ -420,7 +427,7 @@ def test_pyproxy_mixins1(selenium):
         let result = {};
         for(let [name, x] of Object.entries(name_proxy)){
             let impls = {};
-            for(let [name, key] of [
+            for (let [name, key] of [
                 ["then", "then"],
                 ["catch", "catch"],
                 ["finally_", "finally"],
@@ -429,10 +436,11 @@ def test_pyproxy_mixins1(selenium):
             ]){
                 impls[name] = key in x;
             }
-            for(let name of ["PyAwaitable", "PyIterable", "PyIterator"]){
+            for (let name of ["PyAwaitable", "PyIterable", "PyIterator"]){
                 impls[name] = x instanceof pyodide.ffi[name];
             }
             result[name] = impls;
+            assert(() => !("asJsJson" in x))
             x.destroy();
         }
         return result;
@@ -493,6 +501,7 @@ def test_pyproxy_mixins2(selenium):
         assert(() => d.$get.type === "builtin_function_or_method");
         assert(() => d.get.type === undefined);
         assert(() => d.set.type === undefined);
+        assert(() => "asJsJson" in d);
         d.destroy();
         """
     )
@@ -670,6 +679,7 @@ def test_pyproxy_mixins5(selenium):
             assert(() => !("length" in Test));
             assert(() => t.length === 9);
             assert(() => t instanceof pyodide.ffi.PyProxyWithLength);
+            assert(() => !("asJsJson" in t));
             assertThrows(() => {t.length = 10}, "TypeError", "");
             assert(() => t.length === 9);
 
@@ -700,6 +710,7 @@ def test_pyproxy_mixins6(selenium):
         assert(() => l instanceof pyodide.ffi.PyProxyWithHas);
         assert(() => l instanceof pyodide.ffi.PyProxyWithGet);
         assert(() => l instanceof pyodide.ffi.PyProxyWithSet);
+        assert(() => "asJsJson" in l);
         l.set(0, 80);
         pyodide.runPython(`
             assert l[0] == 80
@@ -1090,6 +1101,87 @@ def test_pyproxy_call(selenium):
     with pytest.raises(selenium.JavascriptException, match=msg):
         selenium.run_js("f.callKwargs(76, {x : 6})")
 
+    selenium.run_js("f.destroy()")
+
+
+def test_pyproxy_call_relaxed(selenium):
+    selenium.run_js(
+        """
+        pyodide.runPython(`
+            from pyodide.ffi import to_js
+            def f(x=2, y=3):
+                return to_js([x, y])
+        `);
+        self.f = pyodide.globals.get("f");
+        """
+    )
+
+    def assert_call(s, val):
+        res = selenium.run_js(f"return {s};")
+        assert res == val
+
+    assert_call("f.callRelaxed()", [2, 3])
+    assert_call("f.callRelaxed(1)", [1, 3])
+    assert_call("f.callRelaxed(1, 2)", [1, 2])
+    assert_call("f.callRelaxed(1, 2, 3)", [1, 2])
+    assert_call("f.callRelaxed(1, 2, 3, 4)", [1, 2])
+    selenium.run_js("f.destroy()")
+
+
+def test_pyproxy_call_with_options(selenium):
+    selenium.run_js(
+        """
+        pyodide.runPython(`
+            from pyodide.ffi import to_js
+            def f(x=2, y=3):
+                return to_js([x, y])
+        `);
+        self.f = pyodide.globals.get("f");
+        """
+    )
+
+    def assert_call(s, val):
+        res = selenium.run_js(f"return {s};")
+        assert res == val
+
+    assert_call("f.callWithOptions({})", [2, 3])
+    assert_call("f.callWithOptions({}, 7)", [7, 3])
+    assert_call("f.callWithOptions({}, 7, 9)", [7, 9])
+    msg = r"TypeError: f\(\) takes from 0 to 2 positional arguments but 3 were given"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run_js("f.callWithOptions({}, 7, 9, 10)")
+
+    assert_call("f.callWithOptions({relaxed: true})", [2, 3])
+    assert_call("f.callWithOptions({relaxed: true}, 7)", [7, 3])
+    assert_call("f.callWithOptions({relaxed: true}, 7, 9)", [7, 9])
+    assert_call("f.callWithOptions({relaxed: true}, 7, 9, 10)", [7, 9])
+
+    msg = "callWithOptions with 'kwargs: true' requires at least one argument"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run_js("f.callWithOptions({kwargs: true})")
+
+    msg = "callWithOptions with 'kwargs: true' requires at least one argument"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run_js("f.callWithOptions({kwargs: true, relaxed: true})")
+
+    msg = "kwargs argument is not an object"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run_js("f.callWithOptions({kwargs: true}, 7, 9, 10)")
+
+    msg = "kwargs argument is not an object"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run_js("f.callWithOptions({kwargs: true, relaxed: true}, 7, 9, 10)")
+
+    msg = r"TypeError: f\(\) takes from 0 to 2 positional arguments but 3 were given"
+    with pytest.raises(selenium.JavascriptException, match=msg):
+        selenium.run_js("f.callWithOptions({kwargs: true}, 7, 9, 10, {})")
+
+    assert_call(
+        "f.callWithOptions({kwargs: true, relaxed: true}, 7, 9, 10, {})", [7, 9]
+    )
+
+    assert_call("f.callWithOptions({kwargs: true}, {x: 9})", [9, 3])
+    assert_call("f.callWithOptions({kwargs: true, relaxed: true}, {x: 9})", [9, 3])
     selenium.run_js("f.destroy()")
 
 
@@ -2354,3 +2446,158 @@ def test_automatic_coroutine_scheduling(selenium):
         """
     )
     assert res == [3, 4, 6]
+
+
+@run_in_pyodide
+def test_stringify_sequence(selenium):
+    from json import loads
+
+    from js import JSON
+
+    l = [1, 2, 3]
+
+    assert loads(JSON.stringify(l)) == l  # type:ignore[arg-type]
+    assert loads(JSON.stringify(tuple(l))) == l  # type:ignore[arg-type]
+
+
+@run_in_pyodide
+def test_as_json_adaptor_heritability1(selenium):
+    from pyodide.code import run_js
+
+    o = [1, 2, {"a": 3, "b": {"c": 4, "d": [1, {"c": 2}]}}]
+
+    f = run_js(
+        """
+        (o, l) =>  {
+            const o2 = o.asJsJson();
+            return l.reduce((x, y) => x[y], o2);
+        }
+        """
+    )
+
+    assert f(o, [0]) == 1
+    assert f(o, [1]) == 2
+    assert f(o, [2, "a"]) == 3
+    assert f(o, [2, "b", "c"]) == 4
+    assert f(o, [2, "b", "d", 0]) == 1
+    assert f(o, [2, "b", "d", 1, "c"]) == 2
+
+
+@run_in_pyodide
+def test_as_json_adaptor_heritability2(selenium):
+    from pyodide.code import run_js
+
+    class T1:
+        a = {"x": 2}
+
+    class T2:
+        a = {"x": 2}
+
+        def __getitem__(self, key):
+            return {"y": 3}
+
+    o = [T1(), T2()]
+
+    f = run_js(
+        """
+        (o) => {
+            const x = o.asJsJson();
+            const x0 = x[0];
+            const x1 = x[1];
+            const x1a = x1.a;
+            return pyodide.toPy({
+                x0,
+                x1,
+                xflags: x.$$flags,
+                x0flags: x0.$$flags,
+                x1flags: x1.$$flags,
+                x0a: x0.a,
+                x1a,
+                x1ay: x1a.y,
+            });
+        }
+        """
+    )
+    res = f(o)
+    assert res["xflags"] & (1 << 16) != 0
+    assert res["x0flags"] & (1 << 15) == 0
+    assert res["x1flags"] & (1 << 15) != 0
+    assert res["x0a"] == {"x": 2}
+    assert res["x1a"] == {"y": 3}
+    assert res["x1ay"] == 3
+
+
+@run_in_pyodide
+def test_as_json_adaptor_ownkeys(selenium):
+    from pyodide.code import run_js
+
+    o = {"c": 7, "x": 99, "z": 29}
+    f = run_js("(o) => Reflect.ownKeys(o.asJsJson())")
+    assert set(f(o)) == set(o.keys())
+
+
+@run_in_pyodide
+def test_as_json_adaptor_get(selenium):
+    from pyodide.code import run_js
+
+    o = {0: "a", "1": "b", "3c": 4}
+
+    assert run_js("(o) => o.asJsJson()['']")(o) is None
+
+    assert run_js("(o) => o.asJsJson()[0]")(o) == "a"
+    assert run_js("(o) => o.asJsJson()['0']")(o) == "a"
+    assert run_js("(o) => o.asJsJson().get(0)")(o) == "a"
+    assert run_js("(o) => o.asJsJson().get('0')")(o) is None
+
+    assert run_js("(o) => o.asJsJson()[1]")(o) == "b"
+    assert run_js("(o) => o.asJsJson()['1']")(o) == "b"
+    assert run_js("(o) => o.asJsJson().get(1)")(o) is None
+    assert run_js("(o) => o.asJsJson().get('1')")(o) == "b"
+
+    assert run_js("(o) => o.asJsJson()['3c']")(o) == 4
+
+
+@run_in_pyodide
+def test_as_json_adaptor_set(selenium):
+    from pyodide.code import run_js
+
+    o = {}  # type:ignore[var-annotated]
+
+    run_js(
+        """
+        (o) => {
+            x = o.asJsJson();
+            x.a = 1;
+            x[2] = 3;
+            x["4"] = 5;
+            x.b5 = "c";
+            x[""] = 6;
+            assert(() => (typeof x[0] == "undefined"));
+        }
+        """
+    )(o)
+
+    assert o == {"a": 1, 2: 3, 4: 5, "b5": "c", "": 6}
+    assert "2" not in o
+    assert "4" not in o
+
+
+@pytest.mark.parametrize(
+    "o",
+    [
+        [7],
+        [[7]],
+        {"c": 7},
+        {"c": [7]},
+        [1, {"c": 2}],
+        [1, 2, {"a": 3, "b": {"c": 4, "d": [1, {"c": 2}]}}],
+    ],
+)
+@run_in_pyodide
+def test_as_json_adaptor_stringify(selenium, o):
+    from json import loads
+
+    from pyodide.code import run_js
+
+    f = run_js("(o) => JSON.stringify(o.asJsJson())")
+    assert loads(f(o)) == o
