@@ -35,6 +35,11 @@ ROOT = pathlib.Path(__file__).resolve().parent.parent
 Target = namedtuple("target", ("file", "pattern", "prerelease"))
 PYTHON_TARGETS = [
     Target(
+        file=ROOT / "Makefile.envs",
+        pattern=build_version_pattern(r"PYODIDE_VERSION \?= {python_version}"),
+        prerelease=True,
+    ),
+    Target(
         file=ROOT / "src/py/pyodide/__init__.py",
         pattern=build_version_pattern('__version__ = "{python_version}"'),
         prerelease=True,
@@ -42,11 +47,6 @@ PYTHON_TARGETS = [
     Target(
         file=ROOT / "src/py/pyproject.toml",
         pattern=build_version_pattern('version = "{python_version}"'),
-        prerelease=True,
-    ),
-    Target(
-        ROOT / "pyodide-build/pyodide_build/__init__.py",
-        pattern=build_version_pattern('__version__ = "{python_version}"'),
         prerelease=True,
     ),
     Target(
@@ -120,7 +120,7 @@ def parse_current_version(target: Target) -> str:
     match = target.pattern.search(content)
 
     if match is None:
-        raise ValueError(f"Unabled to detect version string: {target.file}")
+        raise ValueError(f"Unable to detect version string: {target.file}")
 
     return match.groupdict()["version"]
 
@@ -178,6 +178,11 @@ def parse_args():
     parser.add_argument(
         "--dry-run", action="store_true", help="Don't actually write anything"
     )
+    parser.add_argument(
+        "--check",
+        action="store_true",
+        help="Compare the current contents to the updated contents and fail if it would change anything",
+    )
 
     return parser.parse_args()
 
@@ -210,12 +215,20 @@ def main():
         if new_content is not None:
             update_queue.append((target, new_content))
 
+    if args.check:
+        if update_queue:
+            print("Version update would change files, failing", file=sys.stderr)
+            return 1
+        return 0
     if args.dry_run:
-        return
+        return 0
 
     for target, content in update_queue:
         target.file.write_text(content)
+    return 0
 
 
 if __name__ == "__main__":
-    main()
+    import sys
+
+    sys.exit(main())
